fix(router) consider host for prefix candidates
Include supplementary hostname information when generating prefix match
candidates.

When searching through the list of prefix URI match candidates, confirm
that the current request hostname is in the list of hostnames matchable
on a given prefix. This ensures that a longer matching prefix on that
does not include a hostname match does not preclude matching shorter
prefixes that do include a hostname match.